fix: Configuration and usage of PagedModel
Fixes this log warning :
Serializing PageImpl instances as-is is not supported, meaning that there is no guarantee about the stability of the resulting JSON structure!
For a stable JSON structure, please use Spring Data's PagedModel or Spring HATEOAS and Spring Data's PagedResourcesAssembler.

Source: https://bootify.io/spring-boot/pagination-in-spring-boot-rest-api.html#configuration-and-usage-of-pagedmodel